+++ cmd/core/removebackend.go
@@ -4,11 +4,13 @@
 package core
 
 import (
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"regexp"
+	"strings"
 	"time"
 
 	"github.com/go-git/go-git/v5"
@@ -46,6 +48,13 @@ This command will perform the following actions in each cloned repository specif
 			}
 		}
 
+		metadata, err := loadMetadataRemoveBackend("terraform_config_metadata.json")
+		if err != nil {
+			return fmt.Errorf("error loading metadata: %v. Run tfm core init-repos first", err)
+		}
+
+		_ = metadata
+
 		clonePath := viper.GetString("github_clone_repos_path")
 		branchName := "update-backend-" + time.Now().Format("20060102")
 
@@ -81,6 +90,20 @@ This command will perform the following actions in each cloned repository specif
 func init() {
 	CoreCmd.AddCommand(RemoveBackendCmd)
 	RemoveBackendCmd.Flags().BoolVar(&autoApprove, "autoapprove", false, "Automatically approve the operation without a confirmation prompt")
+}
+
+// Loads the metadata file information for use
+func loadMetadataRemoveBackend(metadataFile string) ([]RepoConfig, error) {
+	var metadata []RepoConfig
+	file, err := ioutil.ReadFile(metadataFile)
+	if err != nil {
+		return nil, fmt.Errorf("error reading metadata file: %v. Run tfm core init-repos first.", err)
+	}
+	err = json.Unmarshal(file, &metadata)
+	if err != nil {
+		return nil, fmt.Errorf("error unmarshaling metadata: %v", err)
+	}
+	return metadata, nil
 }
 
 // detectBackendBlocks checks if there's a backend block in any .tf file within the repo.
@@ -191,59 +214,125 @@ func createBranchIfNeeded(clonePath, branchName string) ([]string, error) {
 }
 
 func removeBackendFromRepos(clonePath string) error {
-	dirs, err := os.ReadDir(clonePath)
+	metadata, err := loadMetadataRemoveBackend("terraform_config_metadata.json")
 	if err != nil {
-		return fmt.Errorf("error reading clone path directories: %v", err)
+		return fmt.Errorf("error loading metadata: %v. Run tfm core init-repos first", err)
 	}
 
 	backendRegexp := regexp.MustCompile(`(?s)backend\s+"[^"]+"\s+\{.*?\}`)
 
-	for _, dir := range dirs {
-		if !dir.IsDir() {
-			continue
-		}
+	for _, repoConfig := range metadata {
+		for _, configPath := range repoConfig.ConfigPaths {
+			fullPath := constructFullPath(clonePath, repoConfig, configPath)
+			// fullPath := ""
+			// if strings.HasPrefix(configPath.Path, repoConfig.RepoName+"/") {
+			// 	// If configPath already includes the repoName, use it directly
+			// 	fullPath = filepath.Join(clonePath, configPath.Path)
+			// } else {
+			// 	// If not, concatenate repoName with configPath
+			// 	fullPath = filepath.Join(clonePath, repoConfig.RepoName, configPath.Path)
+			// }
 
-		repoPath := filepath.Join(clonePath, dir.Name())
-		files, err := ioutil.ReadDir(repoPath)
-		if err != nil {
-			fmt.Printf("Error reading repo directory: %v\n", err)
-			continue
-		}
-
-		repoModified := false
-
-		for _, file := range files {
-			if filepath.Ext(file.Name()) != ".tf" {
-				continue
-			}
-
-			filePath := filepath.Join(repoPath, file.Name())
-			content, err := ioutil.ReadFile(filePath)
-			if err != nil {
-				fmt.Printf("Error reading .tf file: %v\n", err)
-				continue
-			}
-
-			modifiedContent := backendRegexp.ReplaceAll(content, []byte(""))
-
-			if len(modifiedContent) != len(content) {
-				err = ioutil.WriteFile(filePath, modifiedContent, file.Mode())
+			err := filepath.Walk(fullPath, func(path string, info os.FileInfo, err error) error {
 				if err != nil {
-					fmt.Printf("Error writing modified .tf file: %v\n", err)
-					continue
+					return err
 				}
-				fmt.Printf("Removed backend configuration from: %s\n", filePath)
-				repoModified = true
+				if !info.IsDir() && strings.HasSuffix(info.Name(), ".tf") {
+					content, readErr := ioutil.ReadFile(path)
+					if readErr != nil {
+						return readErr
+					}
+					modifiedContent := backendRegexp.ReplaceAll(content, []byte(""))
+					if len(modifiedContent) != len(content) {
+						writeErr := ioutil.WriteFile(path, modifiedContent, info.Mode())
+						if writeErr != nil {
+							return writeErr
+						}
+						fmt.Printf("Removed backend configuration from: %s\n", path)
+					}
+				}
+				return nil
+			})
+			if err != nil {
+				fmt.Printf("Error processing files in %s: %v\n", fullPath, err)
 			}
-		}
-
-		if !repoModified {
-			fmt.Printf("No backend blocks found in: %s\n", repoPath)
 		}
 	}
 
 	return nil
 }
+
+func constructFullPath(clonePath string, repoConfig RepoConfig, configPath ConfigPathInfo) string {
+	// Check if configPath.Path is equivalent to the root of the repository
+	isRootPath := configPath.Path == repoConfig.RepoName || configPath.Path == ""
+	if isRootPath {
+		// If configPath is the root, the fullPath is just the clonePath joined with repoName
+		return filepath.Join(clonePath, repoConfig.RepoName)
+	} else {
+		// For non-root configPaths, ensure we don't duplicate the repoName if it's already included
+		if strings.HasPrefix(configPath.Path, repoConfig.RepoName+"/") {
+			return filepath.Join(clonePath, configPath.Path)
+		} else {
+			return filepath.Join(clonePath, repoConfig.RepoName, configPath.Path)
+		}
+	}
+}
+
+// func removeBackendFromRepos(clonePath string) error {
+// 	dirs, err := os.ReadDir(clonePath)
+// 	if err != nil {
+// 		return fmt.Errorf("error reading clone path directories: %v", err)
+// 	}
+
+// 	backendRegexp := regexp.MustCompile(`(?s)backend\s+"[^"]+"\s+\{.*?\}`)
+
+// 	for _, dir := range dirs {
+// 		if !dir.IsDir() {
+// 			continue
+// 		}
+
+// 		repoPath := filepath.Join(clonePath, dir.Name())
+// 		files, err := ioutil.ReadDir(repoPath)
+// 		if err != nil {
+// 			fmt.Printf("Error reading repo directory: %v\n", err)
+// 			continue
+// 		}
+
+// 		repoModified := false
+
+// 		for _, file := range files {
+// 			if filepath.Ext(file.Name()) != ".tf" {
+// 				continue
+// 			}
+
+// 			filePath := filepath.Join(repoPath, file.Name())
+// 			content, err := ioutil.ReadFile(filePath)
+// 			if err != nil {
+// 				fmt.Printf("Error reading .tf file: %v\n", err)
+// 				continue
+// 			}
+
+// 			modifiedContent := backendRegexp.ReplaceAll(content, []byte(""))
+
+// 			if len(modifiedContent) != len(content) {
+// 				err = ioutil.WriteFile(filePath, modifiedContent, file.Mode())
+// 				if err != nil {
+// 					fmt.Printf("Error writing modified .tf file: %v\n", err)
+// 					continue
+// 				}
+// 				fmt.Printf("Removed backend configuration from: %s\n", filePath)
+// 				repoModified = true
+// 			}
+// 		}
+
+// 		if !repoModified {
+// 			fmt.Printf("No backend blocks found in: %s\n", repoPath)
+// 		}
+// 	}
+
+// 	return nil
+// }
+
 
 func commitChanges(repoPath, branchName string) error {
 	commitMessage := viper.GetString("commit_message")